Struct skylight::oleauto::bstr::BStr

#[repr(transparent)]
pub struct BStr(_);

A BStr allocated using SysAllocString or similar. This type may or may not contain valid UTF16.

Construction

It is suggested to use the TryFrom impls for constructing this struct. To reduce the amount of generic code, some functions may use BStr directly in their parameters rather than something akin to impl TryInto<BStr>. The user should use BStr::new("data") in these cases.

Panics

The methods on the Deref impl BStrRef will panic if the BStr’s length cannot fit in a usize.

Implementations

impl BStr

pub fn new<D>(data: D) -> Self where
D: TryInto<BStr, Error = BStrCreationError>,

Make a new BStr. This is a convenience function for using the TryFrom impls. Use BStr::try_new() for a non-panicking variant of this constructor.

Panics

Panics if a new BStr could not be allocated, if the length cannot be stored in a u32, or if the number of items in the iterator does not match the reported length (assuming an iterator was passed here).

pub fn try_new<D>(data: D) -> Result<Self, BStrCreationError> where
D: TryInto<BStr, Error = BStrCreationError>,

Make a new BStr. This is a convenience function for using the TryFrom impls. Use BStr::new() for a panicking variant of this constructor.

Errors

Returns a BStrCreationError if a new BStr could not be allocated, if the length cannot be stored in a u32, or if the number of items in the iterator does not match the reported length (assuming an iterator was passed here).

pub fn from_wide_iter(
iter: impl Iterator<Item = u16>,
len: usize
) -> Result<Self, BStrCreationError>

Try to make a new BStr from a wide char iterator.

Errors

Returns a BStrCreationError if a new BStr could not be allocated, if the length cannot be stored in a u32, or if the number of items in the iterator does not match the reported length.

pub fn from_wide_slice(slice: &[u16]) -> Result<Self, BStrCreationError>

Try to make a new BStr from a wide char slice.

Errors

Returns a BStrCreationError if a new BStr could not be allocated or if the length cannot be stored in a u32.

pub unsafe fn from_raw(ptr: *mut u16) -> Self

Make a new BStr from a raw BSTR ptr.

Safety

ptr must be a BSTR allocated with SysAllocStringLen or similar.

Panics

Panics if ptr is null.

pub fn into_raw(self) -> *mut u16

Leak this BStr and return the inner pointer.

pub fn as_bstr_ref(&self) -> &BStrRef

Get this BStr as a &BStrRef.

Panics

Panics if the size of the string cannot fit in a usize.

pub fn as_mut_bstr_ref(&mut self) -> &mut BStrRef

Get this BStr as a &mut BStrRef.

Panics

Panics if the size of the string cannot fit in a usize.