Struct skylight::tlhelp32::ProcessEntry

pub struct ProcessEntry(_);

A Process Entry.

Implementations

impl ProcessEntry

pub fn pid(&self) -> u32

Get the PID of this ProcessEntry.

pub fn num_threads(&self) -> u32

Get the number of threads created by this process.

pub fn thread_base_priority(&self) -> i32

Get the thread base priority of this process.

pub fn exe_name_wide_slice(&self) -> &[u16]

Get the exe name as a wide character slice. This may or may not be valid UTF16.

pub fn exe_name(&self) -> OsString

Get the exe name as an OsString. This allocates per call, so cache the result. If you want possibly-malformed utf16 without allocating, use ProcessEntry::exe_name_wide_slice instead.